+++ tests/Pipaslot.Mediator.Http.Tests/Serialization/ContractSerializer_CredibilityTestBase.cs
@@ -11,7 +11,7 @@ namespace Pipaslot.Mediator.Http.Tests.Serialization
     {
         protected Mock<ICredibleProvider> CredibleProviderMock = new();
         protected abstract IContractSerializer CreateSerializer(ICredibleProvider provider);
-        private IContractSerializer CreateSerializer()
+        protected IContractSerializer CreateSerializer()
         {
             return CreateSerializer(CredibleProviderMock.Object);
         }
